+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Argument parsing modifier to parse an array of a given type: `+`
 - Implemented `sll_arg_state_t`, `sll_free_args` and `sll_parse_args_list`
-- Implemented `sll_complex_add`, `sll_complex_conjugate`, `sll_complex_div`, `sll_complex_mult`, `sll_complex_reciprocal` and `sll_complex_sub` ([#262])
+- Implemented `sll_complex_abs`, `sll_complex_add`, `sll_complex_conjugate`, `sll_complex_div`, `sll_complex_mult`, `sll_complex_reciprocal` and `sll_complex_sub` ([#262])
 - Implemented `sll_complex_t` ([#262])
 - Object creation modifier to dereference array elements: `*`